モデルのfieldを一部変更

+++ backend/transaction/models.py
@@ -24,7 +24,7 @@ class Transaction(models.Model):
     user = models.ForeignKey(CustomUser, on_delete=models.CASCADE, related_name='transactions')
     type = models.CharField(max_length=10, choices=[('収入', '収入'), ('支出', '支出')])
     store = models.CharField(max_length=255, null=True, blank=True)
-    total_price = models.DecimalField(max_digits=10,decimal_places=2, default=0)
+    total_price = models.IntegerField(default=0)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
 
@@ -35,8 +35,8 @@ class Item(models.Model):
     uuid = models.UUIDField(default=uuid.uuid4, editable=False, primary_key=True)
     transaction = models.ForeignKey(Transaction, on_delete=models.CASCADE, related_name='items')
     name = models.CharField(max_length=255)
-    price = models.DecimalField(max_digits=10, decimal_places=2)
-    amount = models.DecimalField(max_digits=10,decimal_places=2)
+    price = models.IntegerField()
+    amount = models.PositiveIntegerField()
     category = models.ForeignKey(Category, on_delete=models.SET_NULL, null=True, blank=True)
 
     def __str__(self):
